+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 * API:
     * Add validation for `.whl` filaname. (\#2147)
-
+    * Trim whitespaces in `DatasetCreateV2.zarr_dir` (\#2138).
 
 # 2.10.2