+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 # Unreleased
 
+* API:
+    * Add validation for `.whl` filename (\#2147).
+    * Trim whitespaces in `DatasetCreateV2.zarr_dir` (\#2138).
+
+# 2.10.2
 
 * App:
     * Add `FRACTAL_PIP_CACHE_DIR` configuration variable (\#2141).